refactor(server): extract run-event payload bounding from heartbeat.ts
Moves payload bounding/truncation, inline base64 image redaction, log
chunk compaction and the board progress-summary redactor into
heartbeat-run-event-payload.ts with their size-cap constants, following
the heartbeat-<domain> sibling pattern. heartbeat.ts re-exports the
previously public symbols. No behavior change.
